backend/requirements.minimal.txt
# Core FastAPI and web server
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
python-multipart==0.0.20

# MongoDB driver
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Authentication and security
passlib==1.7.4
bcrypt==4.1.3
PyJWT==2.10.1
cryptography==46.0.2

# OpenAI SDK
openai==1.99.9

# Environment and configuration
python-dotenv==1.1.1

# Email validation
email-validator==2.3.0

# Core dependencies
pydantic==2.12.0
pydantic_core==2.41.1
anyio==4.11.0
idna==3.10
certifi==2025.10.5
sniffio==1.3.1
typing_extensions==4.15.0
annotated-types==0.7.0
